stable_put_toa+aoa

+++ openair1/PHY/NR_TRANSPORT/nr_transport_proto.h
@@ -141,7 +141,7 @@ int nr_rx_pusch(PHY_VARS_gNB *gNB,
                 uint32_t frame,
                 uint8_t slot,
                 unsigned char harq_pid,
-                short *aoaptr);
+                int *aoaptr);
 
 /** \brief This function performs RB extraction (signal and channel estimates) (currently signal only until channel estimation and compensation are implemented)
     @param rxdataF pointer to the received frequency domain signal
